[refactor] 전시회 상세 코드 분리 및 소개 html 렌더링 사용
